hotshot_state_prover::circuit

Function build

pub(crate) fn build<F, P, STIter, BitIter, SigIter>(
    stake_table_entries: STIter,
    signer_bit_vec: BitIter,
    signatures: SigIter,
    lightclient_state: &GenericLightClientState<F>,
    stake_table_state: &GenericStakeTableState<F>,
    stake_table_capacity: usize,
) -> Result<(PlonkCircuit<F>, GenericPublicInput<F>), PlonkError>
where
    F: RescueParameter,
    P: TECurveConfig<BaseField = F>,
    STIter: IntoIterator,
    STIter::Item: Borrow<(VerKey<P>, U256)>,
    STIter::IntoIter: ExactSizeIterator,
    BitIter: IntoIterator,
    BitIter::Item: Borrow<F>,
    BitIter::IntoIter: ExactSizeIterator,
    SigIter: IntoIterator,
    SigIter::Item: Borrow<Signature<P>>,
    SigIter::IntoIter: ExactSizeIterator,

A function that takes as input:

a list of stake table entries (Vec<(SchnorrVerKey, Amount)>)

a bit vector indicates the signers

a list of schnorr signatures of the updated states (Vec<SchnorrSignature>), default if the node doesn’t sign the state

updated light client state ((view_number, block_height, block_comm_root))

the static stake table state (containing 3 commitments to the 3 columns of the stake table and a threshold)

Lengths of input vectors should not exceed the stake_table_capacity. The list of stake table entries, bit indicators and signatures will be padded to the stake_table_capacity. It checks that

the vector that indicates who signed is a bit vector

the signers’ accumulated weight exceeds the quorum threshold

the stake table corresponds to the one committed in the light client state

all Schnorr signatures over the light client state are valid

and returns

A circuit for proof generation

A list of public inputs for verification

A PlonkError if any error happens when building the circuit