hotshot_state_prover

Module legacy

Legacy prover Legacy prover

Re-exports

pub use snark::*;

Modules

circuit

State verifier circuit builder Legacy circuit implementation for verifying light client state update Circuit implementation is copied from commit 5d0baf07cbb27f4ac63df116bc3ff78545554cdb.

mock_ledger

Utilities for test

service

Prover service related functionalities A light client prover service

snark

SNARK proof generation