hotshot_state_prover::legacy

Module service

Prover service related functionalities A light client prover service

Functions

fetch_latest_legacy_state

Get the latest LightClientState and signature bundle from Sequencer network

generate_proof 🔒

is_contract_legacy

Check if the contract is a legacy contract. Returns true if the version check succeeds, false otherwise.

load_proving_key

read_contract_state

Read the following info from the LightClient contract storage on chain

run_prover_once

Run light client state prover once

run_prover_service

Run prover in daemon mode

start_http_server 🔒

submit_state_and_proof

submit the latest finalized state along with a proof to the L1 LightClient contract

sync_state

Sync the light client state from the relay server and submit the proof to the L1 LightClient contract