hotshot_state_prover::service

Function advance_epoch

async fn advance_epoch(
    state: &mut ProverServiceState,
    provider: impl Provider,
    light_client_address: Address,
    cur_st_state: StakeTableState,
    proving_key: &ProvingKey,
    contract_epoch: Option<<SeqTypes as NodeType>::Epoch>,
    target_epoch: Option<<SeqTypes as NodeType>::Epoch>,
) -> Result<StakeTableState, ProverError>

This function will fetch the cross epoch state update information from the sequencer query node and update the light client state in the contract to the target_epoch. In the end, both the locally stored stake table and the contract light client state will correspond to the target_epoch. It returns the final stake table state at the target epoch.