Module hotshot_state_prover::circuit

State verifier circuit builder Circuit implementation for verifying light client state update

Structs

LightClientStateVar

Light client state Variable

StakeTableCommVar

Light client state Variable The stake table commitment is a triple (qc_keys_comm, state_keys_comm, stake_amount_comm). Variable for a stake table commitment

StakeTableEntryVar

Variable for stake table entry

Functions

build 🔒

A function that takes as input:

build_for_preprocessing 🔒

Internal function to build a dummy circuit

u256_to_field 🔒

Lossy conversion of a U256 into a field element.