Function hotshot_state_prover::circuit::build

pub(crate) fn build<F, P, STIter, BitIter, SigIter>(
    stake_table_entries: STIter,
    signer_bit_vec: BitIter,
    signatures: SigIter,
    lightclient_state: &LightClientState<F>,
    threshold: &U256
) -> Result<(PlonkCircuit<F>, PublicInput<F>), PlonkError>where
    F: RescueParameter,
    P: TECurveConfig<BaseField = F>,
    STIter: IntoIterator,
    STIter::Item: Borrow<(VerKey<P>, U256)>,
    STIter::IntoIter: ExactSizeIterator,
    BitIter: IntoIterator,
    BitIter::Item: Borrow<bool>,
    BitIter::IntoIter: ExactSizeIterator,
    SigIter: IntoIterator,
    SigIter::Item: Borrow<Signature<P>>,
    SigIter::IntoIter: ExactSizeIterator,

A function that takes as input:

a list of stake table entries (Vec<(SchnorrVerKey, Amount)>)

a bit vector indicates the signers

a list of schnorr signatures of the updated states (Vec<SchnorrSignature>), default if the node doesn’t sign the state

updated light client state ((view_number, block_height, block_comm_root, fee_ledger_comm, stake_table_comm))

a quorum threshold Lengths of input vectors should not exceed the STAKE_TABLE_CAPACITY. The list of stake table entries, bit indicators and signatures will be padded to the STAKE_TABLE_CAPACITY. It checks that

the signer’s accumulated weight exceeds the quorum threshold

the stake table corresponds to the one committed in the light client state

all signed Schnorr signatures are valid and returns

A circuit for proof generation

A list of public inputs for verification

A PlonkError if any error happens when building the circuit