Module hotshot_state_prover::service

Prover service related functionalities A light client prover service

Structs

PublicHotShotConfig 🔒

PublicNetworkConfig 🔒

StateProverConfig

Configuration/Parameters used for hotshot state prover

Enums

ProverError

Functions

fetch_latest_state

init_stake_table

init_stake_table_from_sequencer 🔒

Initialize the stake table from a sequencer node that is currently providing the HotShot config.

light_client_genesis

light_client_genesis_from_stake_table

load_proving_key

one_honest_threshold

A helper function to compute the quorum threshold given a total amount of stake.

prepare_contract 🔒

prepare a contract interface ready to be read from or written to

read_contract_state

get the finalizedState from the LightClient contract storage on L1

run_prover_once

Run light client state prover once

run_prover_service

run_prover_service_with_stake_table

start_http_server 🔒

submit_state_and_proof

submit the latest finalized state along with a proof to the L1 LightClient contract

sync_state

Type Aliases

F 🔒

SignerWallet

A wallet with local signer and connected to network via http