hotshot_state_prover

Module service

Prover service related functionalities A light client prover service

Structs

ProverServiceState

StateProverConfig

Configuration/Parameters used for hotshot state prover

Enums

ProverError

Functions

advance_epoch 🔒

This function will fetch the cross epoch state update information from the sequencer query node and update the light client state in the contract to the target_epoch. In the end, both the locally stored stake table and the contract light client state will correspond to the target_epoch. It returns the final stake table state at the target epoch.

fetch_epoch_state_from_sequencer 🔒

fetch_latest_state

Get the latest LightClientState and signature bundle from Sequencer network

generate_proof 🔒

load_proving_key

read_contract_state

Read the following info from the LightClient contract storage on chain

run_prover_once

Run light client state prover once

run_prover_service

start_http_server 🔒

submit_state_and_proof

submit the latest finalized state along with a proof to the L1 LightClient contract

sync_state

Sync the light client state from the relay server and submit the proof to the L1 LightClient contract