Function hotshot_state_prover::snark::generate_state_update_proof

pub fn generate_state_update_proof<STIter, R, BitIter, SigIter>(
    rng: &mut R,
    pk: &ProvingKey,
    stake_table_entries: STIter,
    signer_bit_vec: BitIter,
    signatures: SigIter,
    lightclient_state: &LightClientState,
    threshold: &U256,
    stake_table_capacity: usize,
) -> Result<(Proof, PublicInput), PlonkError>
where
    STIter: IntoIterator,
    STIter::Item: Borrow<(StateVerKey, U256)>,
    STIter::IntoIter: ExactSizeIterator,
    R: CryptoRng + RngCore,
    BitIter: IntoIterator,
    BitIter::Item: Borrow<bool>,
    BitIter::IntoIter: ExactSizeIterator,
    SigIter: IntoIterator,
    SigIter::Item: Borrow<Signature<EdwardsConfig>>,
    SigIter::IntoIter: ExactSizeIterator,

Given a proving key and

a list of stake table entries (Vec<(BLSVerKey, Amount, SchnorrVerKey)>)

a list of schnorr signatures of the updated states (Vec<SchnorrSignature>), default if the node doesn’t sign the state

updated light client state ((view_number, block_height, block_comm_root, fee_ledger_comm, stake_table_comm))

a bit vector indicates the signers

a quorum threshold

Returns error or a pair (proof, public_inputs) asserting that

the signer’s accumulated weight exceeds the quorum threshold

the stake table corresponds to the one committed in the light client state

all schnorr signatures over the light client state are valid

Errors

Errors if unable to generate proof

Panics

if the stake table is not up to date